+++ Example/AEImageExample/AppDelegate.swift
@@ -1,6 +1,6 @@
 //
 //  AppDelegate.swift
-//  AEImageDemo
+//  AEImageExample
 //
 //  Created by Marko Tadic on 9/17/16.
 //  Copyright © 2016 AE. All rights reserved.
@@ -10,11 +10,5 @@ import UIKit
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
-
     var window: UIWindow?
-
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
-        return true
-    }
-
 }