+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## Version 2.1.0
+
+- Improvements and refactoring
+- Removed AE prefix from type names
+- Enabled framework to be used in app extensions
+- Improved inline documentation
+
+## Version 2.0.0
+
+- Major refactoring and improvements
+- Motion logic improvements
+
+## Version 1.2.0
+
+- Added infinite scroll effect
+- Bumped deployment target to iOS 9.0
+
+## Version 1.1.0
+
+- Update for Xcode 8.1 Beta (8T47)
+
+## Version 1.0.0
+
+- Published on Github
+
 ## Version 0.1.0
 
 - Initial version